+++ packages/article-list/__tests__/utils.js
@@ -19,3 +19,16 @@ export const omitNative = new Set([
 ]);
 
 export const omitWeb = new Set(["className", "data-testid", "style"]);
+
+const config = {
+  get (renderKey) {
+    return "www.thetimes.co.uk";
+  }
+};
+
+export const makeUrl = (slug, shortIdentifier) => {
+  if (process.env.NODE_ENV === 'local' && !process.env.IS_E2E_CI) {
+      return `${config.get('render:host')}/${config.get('render:port')}/article/${slug}-${shortIdentifier}`;
+  }
+  return `${config.get('render:host')}/article/${slug}-${shortIdentifier}`;
+};
